requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.2
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
filelock==3.0.12
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.0.2
promise==2.3
pycodestyle==2.7.0
pyfingerprint==1.5
PyJWT==1.7.1
pyserial==3.5
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uuid==1.30
virtualenv==20.6.0
virtualenv-clone==0.5.5
